db: follow the schema to bigint GENERATED ALWAYS AS IDENTITY
The SQL files moved to bigint identity keys and bigint foreign keys. The
models said Integer and SERIAL, so the two had diverged; this closes it.

Two details that had to be mirrored rather than approximated.

Identity(always=True), not autoincrement and not a SERIAL default. SERIAL,
no column default, and they behave differently: only ALWAYS refuses an
explicit id. An id written past the sequence leaves it behind the data, and
the next generated value collides long after whatever wrote it has been
forgotten. Verified against Postgres -- an explicit id is rejected with
GeneratedAlways, and OVERRIDING SYSTEM VALUE still works for an importer
that must preserve ids.

with_variant(Integer, "sqlite") on every key and foreign key. SQLite
auto-assigns only for INTEGER PRIMARY KEY, which is an alias for the rowid;
a BIGINT primary key there is an ordinary column and an insert omitting it
fails. Without the variant the entire fast suite would have gone with it.

The schema diff now compares identity too, because the earlier version would
have passed SERIAL against GENERATED ALWAYS without noticing -- the one
difference reflection does not surface as a type or a default. Both schemas
built against real Postgres and diffed clean.

doc/architecture.md said bigserial, which was never what the models did and
is no longer what the SQL does either.

+++ db/schema/create/bookmark_tag.sql
@@ -6,8 +6,9 @@
 --
 
 CREATE TABLE bookmark_tag (
-    user_bookmark_id  integer      NOT NULL,
-    tag_id            integer      NOT NULL,
+    user_bookmark_id  bigint      NOT NULL,
+    tag_id            bigint      NOT NULL,
+
     source            tag_source   NOT NULL DEFAULT 'user',
     -- NULL for a tag a person chose. Set for a machine suggestion, so the UI can show it
     -- as provisional and a bulk accept or reject is a query rather than a migration.
